+++ lib/ecto/adapters/postgres.ex
@@ -60,16 +60,10 @@ defmodule Ecto.Adapters.Postgres do
 
   def create(repo, entity) do
     module      = elem(entity, 0)
-    primary_key = module.__entity__(:primary_key)
-    pk_value    = entity.primary_key
 
-    returning = module.__entity__(:entity_kw, entity)
+    returning = module.__entity__(:keywords, entity)
       |> Enum.filter(fn { _, val } -> val == nil end)
       |> Keyword.keys
-
-    if primary_key && !pk_value do
-      returning = [primary_key] ++ returning
-    end
 
     case query(repo, SQL.insert(entity, returning)) do
       Postgrex.Result[rows: [values]] ->
@@ -214,15 +208,19 @@ defmodule Ecto.Adapters.Postgres do
       Worker.commit!(worker)
       { :ok, value }
     catch
-      :throw, :ecto_rollback ->
+      :throw, { :ecto_rollback, value } ->
         Worker.rollback!(worker)
-        :error
+        { :error, value }
       type, term ->
         Worker.rollback!(worker)
         :erlang.raise(type, term, System.stacktrace)
     after
       checkin_worker(repo)
     end
+  end
+
+  def rollback(_repo, value) do
+    throw { :ecto_rollback, value }
   end
 
   defp use_worker(repo, fun) do