+++ lib/ecto/associations/has_one.ex
@@ -1,9 +1,35 @@
-defrecord Ecto.Reflections.HasOne, [ :field, :owner, :associated,
-  :key, :assoc_key ]
+defrecord Ecto.Reflections.HasOne, [:field, :owner, :associated, :key, :assoc_key] do
+  @moduledoc """
+  The reflection record for a `has_one` association. Its fields are:
+
+  * `field` - The name of the association field on the entity;
+  * `owner` - The model where the association was defined;
+  * `associated` - The model that is associated;
+  * `key` - The key on the `owner` model used for the association;
+  * `assoc_key` - The key on the `associated` model used for the association;
+  """
+end
 
 defmodule Ecto.Associations.HasOne do
   @moduledoc """
   A has_one association.
+
+  ## Reflection
+
+  Any association module will generate the `__assoc__` function that can be
+  used for runtime introspection of the association.
+
+  * `__assoc__(:loaded, assoc)` - Returns the loaded entities or `:not_loaded`;
+  * `__assoc__(:loaded, value, assoc)` - Sets the loaded entities;
+  * `__assoc__(:target, assoc)` - Returns the entity where the association was
+                                  defined;
+  * `__assoc__(:name, assoc)` - Returns the name of the association field on the
+                                entity;
+  * `__assoc__(:primary_key, assoc)` - Returns the primary key (used when
+                                       creating a an entity with `new/2`);
+  * `__assoc__(:primary_key, value, assoc)` - Sets the primary key;
+  * `__assoc__(:new, name, target)` - Creates a new association with the given
+                                      name and target;
   """
 
   alias Ecto.Reflections.HasOne, as: Refl
@@ -50,7 +76,10 @@ defmodule Ecto.Associations.HasOne do
       assoc([{ unquote(field), var }]) = record
       var
     end
+  end
 
+  @doc false
+  Enum.each [:loaded, :primary_key], fn field ->
     def __assoc__(unquote(field), value, record) do
       assoc(record, [{ unquote(field), value }])
     end